+++ contracts/src/finance/vesting_wallet.rs
@@ -40,6 +40,7 @@ use stylus_sdk::{
 use crate::{
     access::ownable::{self, IOwnable, Ownable},
     token::erc20::utils::safe_erc20::{self, ISafeErc20, SafeErc20},
+    utils::math::storage::AddAssignChecked,
 };
 
 #[cfg_attr(coverage_nightly, coverage(off))]
@@ -421,12 +422,10 @@ impl IVestingWallet for VestingWallet {
     fn release_eth(&mut self) -> Result<(), Self::Error> {
         let amount = self.releasable_eth();
 
-        let released = self
-            ._released
-            .get()
-            .checked_add(amount)
-            .expect("total released should not exceed `U256::MAX`");
-        self._released.set(released);
+        self._released.add_assign_checked(
+            amount,
+            "total released should not exceed `U256::MAX`",
+        );
 
         let owner = self.ownable.owner();
 
@@ -442,12 +441,10 @@ impl IVestingWallet for VestingWallet {
         let amount = self.releasable_erc20(token)?;
         let owner = self.ownable.owner();
 
-        let released = self
-            ._erc20_released
-            .get(token)
-            .checked_add(amount)
-            .expect("total released should not exceed `U256::MAX`");
-        self._erc20_released.setter(token).set(released);
+        self._erc20_released.setter(token).add_assign_checked(
+            amount,
+            "total released should not exceed `U256::MAX`",
+        );
 
         self.safe_erc20.safe_transfer(token, owner, amount)?;