+++ contracts/src/utils/math/storage/checked.rs
@@ -3,7 +3,7 @@ use alloy_primitives::Uint;
 use stylus_sdk::storage::StorageUint;
 
 /// Adds value and assign the result to `self`, panicking on overflow.
-pub trait AddAssignChecked<T> {
+pub(crate) trait AddAssignChecked<T> {
     /// Adds `rhs` and assign the result to `self`, panicking on overflow.
     fn add_assign_checked(&mut self, rhs: T, msg: &str);
 }
@@ -13,21 +13,6 @@ impl<const B: usize, const L: usize> AddAssignChecked<Uint<B, L>>
 {
     fn add_assign_checked(&mut self, rhs: Uint<B, L>, msg: &str) {
         let new_balance = self.get().checked_add(rhs).expect(msg);
-        self.set(new_balance);
-    }
-}
-
-/// Subtract value and assign the result to `self`, panicking on overflow.
-pub trait SubAssignChecked<T> {
-    /// Subtract `rhs` and assign the result to `self`, panicking on overflow.
-    fn sub_assign_checked(&mut self, rhs: T, msg: &str);
-}
-
-impl<const B: usize, const L: usize> SubAssignChecked<Uint<B, L>>
-    for StorageUint<B, L>
-{
-    fn sub_assign_checked(&mut self, rhs: Uint<B, L>, msg: &str) {
-        let new_balance = self.get().checked_sub(rhs).expect(msg);
         self.set(new_balance);
     }
 }